refactor: AlertManager 패키지 위치 수정
- AlertManager 인터페이스의 위치를 도메인에서 애플리케이션으로 변경
