+++ src/main/java/com/atwoz/alert/infrastructure/FirebaseAlertManager.java
@@ -1,7 +1,7 @@
 package com.atwoz.alert.infrastructure;
 
 import com.atwoz.alert.domain.Alert;
-import com.atwoz.alert.domain.AlertManager;
+import com.atwoz.alert.application.AlertManager;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.Message;
 import com.google.firebase.messaging.Notification;
